requirements.txt
requests==2.23.0
botbuilder-integration-aiohttp>=4.14.0
langchain
openai
google-api-python-client
google-search-results
wikipedia
python-dotenv
faiss-cpu
tiktoken
pika
O365